// src/app/core/interfaces/Ejercitario.ts
export interface Ejercitario {
  id?: number;
  tipoDeEjercitario?: string;
  nombreDeEjercitario?: string;
  instruccionPrincipalEjercitario?: string;
  variaciones?: null;
  duracion?: number;
  instruccionesParticipantes?: string;
  urlEjercitario?: string;
  nivel?: string;
  categoria?: null;
  sector?: null;
  competencia_id?: number;
  progreso?: number;
}